+++ tests/testthat.R
@@ -22,7 +22,7 @@ test_file("testthat/test_conversion.R")
 test_file("testthat/test_vcfR_methods.R")
 
 test_file("testthat/test_addID.R")
-
+test_file("testthat/test_maf.R")
 
 #test_file("testthat/test_rank_variants.R")
 #test_file("testthat/test_var_window.R")