+++ tests/testthat/test_maf.R
@@ -1,0 +1,18 @@
+
+
+
+library(vcfR)
+context("maf")
+
+data("vcfR_example")
+
+
+test_that("maf works on vcfR",{
+  my.maf <- maf(vcf)
+
+  expect_equal( ncol(my.maf), 4 )
+})
+
+
+
+